style: Move category entity and dto to category.entity package and category.dto package
