+++ src/main/java/gift/category/CategoryRepository.java
@@ -1,5 +1,6 @@
 package gift.category;
 
+import gift.category.entity.Category;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
