+++ src/main/java/gift/category/CategoryService.java
@@ -1,5 +1,6 @@
 package gift.category;
 
+import gift.category.entity.Category;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
