+++ src/main/java/gift/category/dto/CategoryDTO.java
@@ -1,4 +1,4 @@
-package gift.category;
+package gift.category.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.Objects;
