+++ src/main/java/gift/category/entity/Category.java
@@ -1,4 +1,4 @@
-package gift.category;
+package gift.category.entity;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.persistence.Column;
