+++ src/main/java/gift/product/dto/ProductRequestDTO.java
@@ -1,6 +1,6 @@
 package gift.product.dto;
 
-import gift.category.CategoryDTO;
+import gift.category.dto.CategoryDTO;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.Objects;
