+++ src/test/java/gift/product/ProductControllerTest.java
@@ -15,8 +15,8 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import gift.category.Category;
-import gift.category.CategoryDTO;
+import gift.category.entity.Category;
+import gift.category.dto.CategoryDTO;
 import gift.product.dto.ProductRequestDTO;
 import gift.product.dto.ProductResponseDTO;
 import gift.token.JwtProvider;
