+++ src/test/java/gift/product/ProductRepositoryTest.java
@@ -3,7 +3,7 @@ package gift.product;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import gift.category.Category;
+import gift.category.entity.Category;
 import gift.category.CategoryRepository;
 import gift.product.entity.Product;
 import java.util.ArrayList;
